requirements.txt
watchdog>=3.0.0
Pillow>=10.0.0
google-generativeai>=0.3.0
openai>=1.0.0
anthropic>=0.18.0
python-dotenv>=1.0.0

